Development/level3/requirements.txt
aiofiles==23.2.1
anyio==3.7.1
certifi==2024.6.2
chardet==5.2.0
fastapi==0.100.0
groq==0.10.0
h11==0.14.0
httpcore==1.*
httpx==0.27.0
idna==3.4
pydantic==2.7.4
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions>=4.7.0
uvicorn==0.29.0